% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdca_package.R
\name{decline_fit.hyperbolic_fit}
\alias{decline_fit.hyperbolic_fit}
\title{S3 method for class 'decline_fit'}
\usage{
\method{decline_fit}{hyperbolic_fit}(decline_fit_lst, time_lst)
}
\arguments{
\item{decline_fit_lst}{a list object of class 'decline_fit'}

\item{time_lst}{a list object of class 'time'}
}
\value{
a list of estimates for the parameters of the Arps hyperbolic model
}
\description{
Return a list of estimated parameters for the Arps hyperbolic decline model
}
