% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdca_package.R
\name{decline_time}
\alias{decline_time}
\title{Arps time object}
\usage{
decline_time(x, unit = "day")
}
\arguments{
\item{x}{a vector of times or a daily sequence of dates}

\item{unit}{time/date unit of vector x}
}
\value{
a list of class 'time' with all the required parameters for the decline_predict() S3 methods
}
\description{
Create an object of class 'time'
}
\examples{
decline_time_1 <- decline_time(c(1:730), unit = "day")

decline_time_1

decline_time_2 <- decline_time(c(1:240), unit = "month")

decline_time_2

decline_time_3 <- decline_time(c(1:15), unit = "year")

decline_time_3

decline_time_4 <- decline_time(seq(as.Date("2020/1/1"), by = "days",
length.out = 360), unit = "date")

decline_time_4
}
