% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decline_fit_interface.R
\docType{package}
\name{Rdca-package}
\alias{Rdca}
\alias{Rdca-package}
\title{Rdca: DCA Tools for Decline Rate Analysis and EUR Forecast}
\description{
Arps Decline Curve Analysis (DCA) models for production rate, cumulative production, nominal decline rate, the derivative of loss-ratio, and estimated ultimate recovery (EUR) predictions for oil and gas wells. Arps, J. J. (1945) <doi:10.2118/945228-G>. Robertson, S. (1988) <https://www.onepetro.org/general/SPE-18731-MS>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://susaenergy.github.io/Rdca_ws/}
}

}
\author{
\strong{Maintainer}: Farshad Tabasinejad \email{farshad.tabasinejad@susaenergy.com}

}
\keyword{internal}
